requirements_build.txt
# Requirements for EMG Plotter GUI and PyInstaller build
PyQt6>=6.4.0
pandas>=1.3.0
matplotlib>=3.5.0
numpy>=1.20.0,<2.0.0
Pillow>=8.0.0
pyinstaller>=5.0.0
